Abstract duplicate test for data store `get_status`
