+++ tests/phpunit/jobstore/ActionScheduler_DBStore_Test.php
@@ -446,20 +446,4 @@ class ActionScheduler_DBStore_Test extends AbstractStoreTest {
 		$this->assertEquals( (int) ( $now->format( 'U' ) ) + HOUR_IN_SECONDS, $store->get_date( $new_action_id )->format( 'U' ) );
 	}
 
-	public function test_get_status() {
-		$time = as_get_datetime_object('-10 minutes');
-		$schedule = new ActionScheduler_IntervalSchedule($time, HOUR_IN_SECONDS);
-		$action = new ActionScheduler_Action('my_hook', array(), $schedule);
-		$store = new ActionScheduler_DBStore();
-		$action_id = $store->save_action($action);
-
-		$this->assertEquals( ActionScheduler_Store::STATUS_PENDING, $store->get_status( $action_id ) );
-
-		$store->mark_complete( $action_id );
-		$this->assertEquals( ActionScheduler_Store::STATUS_COMPLETE, $store->get_status( $action_id ) );
-
-		$store->mark_failure( $action_id );
-		$this->assertEquals( ActionScheduler_Store::STATUS_FAILED, $store->get_status( $action_id ) );
-	}
-
 }
